# packages/sdk/openapi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdottaUnGatto API",
    "version": "1.0.0",
    "description": "OpenAPI source for the v1 HTTP contract used by generated SDK types."
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Current authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "operationId": "updateCurrentUserPreferences",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/profile": {
      "get": {
        "operationId": "getCurrentUserProfile",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "operationId": "updateCurrentUserProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/avatar": {
      "post": {
        "operationId": "upsertCurrentUserAvatar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAvatarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUserAvatar",
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/consents": {
      "get": {
        "operationId": "getCurrentUserConsents",
        "responses": {
          "200": {
            "description": "Current consent states",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "operationId": "updateCurrentUserConsents",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsentsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated consent states",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/favorites": {
      "get": {
        "operationId": "getCurrentUserFavorites",
        "responses": {
          "200": {
            "description": "Current user favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/favorites/{listingId}": {
      "put": {
        "operationId": "addCurrentUserFavorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUserFavorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/linked-identities": {
      "get": {
        "operationId": "getCurrentUserLinkedIdentities",
        "responses": {
          "200": {
            "description": "Linked external identities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIdentitiesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/linked-identities/{provider}/start": {
      "post": {
        "operationId": "startCurrentUserLinkedIdentity",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderAlias"
          }
        ],
        "responses": {
          "200": {
            "description": "Identity linking start information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartLinkedIdentityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/linked-identities/{provider}": {
      "delete": {
        "operationId": "deleteCurrentUserLinkedIdentity",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderAlias"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated linked identities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIdentitiesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/users/me/sessions": {
      "get": {
        "operationId": "getCurrentUserSessions",
        "responses": {
          "200": {
            "description": "Active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/sessions/{sessionId}": {
      "delete": {
        "operationId": "deleteCurrentUserSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ListingId": {
        "name": "listingId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9]\\\\d*$"
        }
      },
      "ProviderAlias": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9_-]{0,62}$"
        }
      },
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": ["user", "moderator", "admin"]
      },
      "IdentityProvider": {
        "type": "string",
        "enum": ["dev-header", "keycloak"]
      },
...
      "UpdatePreferencesRequest": {
        "type": "object",
        "required": ["messageEmailNotificationsEnabled"],
        "additionalProperties": false,
        "properties": {
          "messageEmailNotificationsEnabled": {
            "type": "boolean"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "provider",
          "providerSubject",
          "email",
          "emailVerified",
          "roles",
          "preferences",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/IdentityProvider"
          },
          "providerSubject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "required": ["user"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "required": ["profile"],
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true,
            "maxLength": 80
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "maxLength": 80
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "phoneE164": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          },
          "city": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "province": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "maxLength": 800
          }
        }
      },
      "UpsertAvatarRequest": {
        "type": "object",
        "required": ["mimeType", "contentBase64"],
        "additionalProperties": false,
        "properties": {
          "mimeType": {
            "type": "string",
            "maxLength": 120
          },
          "contentBase64": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true,
            "maxLength": 180
          }
        }
      },
...
      "ConsentsResponse": {
        "type": "object",
        "required": ["consents"],
        "properties": {
          "consents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserConsent"
            }
          }
        }
      },
...
      "UpdateConsentsRequest": {
        "type": "object",
        "required": ["consents"],
        "additionalProperties": false,
        "properties": {
          "consents": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": {
              "$ref": "#/components/schemas/ConsentUpdate"
            }
          }
        }
      },
...
      "FavoritesResponse": {
        "type": "object",
        "required": ["favorites"],
        "properties": {
          "favorites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserFavorite"
            }
          }
        }
      },
...
      "LinkedIdentitiesResponse": {
        "type": "object",
        "required": ["linkedIdentities"],
        "properties": {
          "linkedIdentities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkedIdentity"
            }
          }
        }
      },
      "StartLinkedIdentityResponse": {
        "type": "object",
        "required": ["provider", "redirectUrl"],
        "properties": {
          "provider": {
            "type": "string"
          },
          "redirectUrl": {
            "type": "string"
          }
        }
      },
...
      "SessionsResponse": {
        "type": "object",
        "required": ["sessions"],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSession"
            }
          }
        }
      }
    }
  }
}